move out not main functions
